Moved miniz to its own library. Added license info
