+++ lib/LLVMBuild.txt
@@ -45,6 +45,7 @@ subdirectories =
  DxilDia
  DxrFallback
  DxilRootSignature
+ Miniz
 
 ; HLSL Change: remove LibDriver, LineEditor, add HLSL, DxrtFallback, DXIL, DxilContainer, DxilDia, DxilPIXPasses, DxilRootSignature
