+++ lib/Miniz/CMakeLists.txt
@@ -1,0 +1,6 @@
+# Copyright (C) Microsoft Corporation. All rights reserved.
+# This file is distributed under the University of Illinois Open Source License. See LICENSE.TXT for details.
+
+add_llvm_library(LLVMMiniz
+  miniz.c
+)
