+++ lib/Miniz/miniz.c
@@ -24,7 +24,7 @@
  *
  **************************************************************************/
 
-#include  "miniz.h"
+#include  "miniz/miniz.h" // HLSL Change
 
 typedef unsigned char mz_validate_uint16[sizeof(mz_uint16) == 2 ? 1 : -1];
 typedef unsigned char mz_validate_uint32[sizeof(mz_uint32) == 4 ? 1 : -1];
